Extract document preview component
